chore: consolidate benchmark scripts and remove obsolete files
